[presto] Disable hostname verification for client-coordinator communication
This is to allow end users with misconfigured certs to ignore SSL hostname
verification when using a presto-client based client (our attention is on
presto-jdbc) to still authenticate and connect.
One would expect this to be needed if they didn't configure SSL, in which
case our presto containers will create a self-signed cert and use that to
SSL-ify.